+++ vaultron.tf
@@ -178,7 +178,6 @@ module "consul_cluster" {
   consul_oss_instance_count    = "${var.consul_oss_instance_count}"
   consul_version               = "${var.consul_version}"
   datacenter_name              = "${var.datacenter_name}"
-  statsd_ip                    = "${module.telemetry.statsd_graphite_ip}"
   use_consul_oss               = "${var.use_consul_oss}"
 }
 
@@ -188,7 +187,6 @@ module "vaultron" {
   consul_server_ips            = ["${module.consul_cluster.consul_oss_server_ips}"]
   consul_client_ips            = ["${module.consul_cluster.consul_oss_client_ips}"]
   disable_clustering           = "${var.disable_clustering}"
-  statsd_ip                    = "${module.telemetry.statsd_graphite_ip}"
   use_vault_oss                = "${var.use_vault_oss}"
   vault_cluster_name           = "${var.vault_cluster_name}"
   vault_custom_config_template = "${var.vault_custom_config_template}"